Add files via upload
I am analyzing a PCAP (Packet Capture) file in Wireshark, and you've filtered the traffic to show HTTP requests.

Packet Information:

The communication is between two IP addresses: 192.168.1.2 (destination) and 192.168.1.3 (source).
The source (192.168.1.3) sent an HTTP GET request to retrieve the file /flag.txt from the destination server.
The server responded with HTTP 200 OK, indicating that the request was successful.
Flag Retrieval:

In the HTTP response, the server returned the contents of the file /flag.txt. You can see the flag in the "File Data" section of the packet:
HSCTF{http1suns4f3}.
This shows that you successfully captured the flag from the HTTP traffic in the PCAP file by analyzing the HTTP GET request and response. The flag is HSCTF{http1suns4f3}.